extract viewUrl to Utils file
