+++ src/tools/web/utils/viewUrlUtils.test.ts
@@ -1,0 +1,12 @@
+import { constructViewWebUrl } from './viewUrlUtils.js';
+
+describe('constructViewWebUrl', () => {
+  it('constructs correct URL from server, site, and contentUrl', () => {
+    const result = constructViewWebUrl(
+      'https://tableau.example.com',
+      'my-site',
+      'workbook/sheets/Sheet1',
+    );
+    expect(result).toBe('https://tableau.example.com/#/site/my-site/views/workbook/Sheet1');
+  });
+});
